Modify :移除討論區所有dao、dao_interface
