+++ src/main/java/com/foodietime/favoritepost/model/FavoritePostVO.java
@@ -12,9 +12,9 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "FAVORITE_POST")
 @IdClass(FavoritePostPK.class)
-public class FavoritePostVO implements Serializable{
-	
-	@Id	
+public class FavoritePostVO implements Serializable {
+
+	@Id
 	@ManyToOne
 	@JoinColumn(name = "POST_ID", referencedColumnName = "POST_ID")
 	private PostVO postId;
@@ -22,11 +22,14 @@ public class FavoritePostVO implements Serializable{
 	@ManyToOne
 	@JoinColumn(name = "MEM_ID", referencedColumnName = "MEM_ID")
 	private MemberVO memId;
-	
+
 	public FavoritePostVO(PostVO postId, MemberVO memId) {
 		super();
 		this.postId = postId;
 		this.memId = memId;
+	}
+
+	public FavoritePostVO() {
 	}
 
 	public PostVO getPostId() {
@@ -44,7 +47,5 @@ public class FavoritePostVO implements Serializable{
 	public void setMemId(MemberVO memId) {
 		this.memId = memId;
 	}
-
-	
 
 }
