+++ src/main/java/com/foodietime/message/model/MessageRepository.java
@@ -1,0 +1,9 @@
+package com.foodietime.message.model;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MessageRepository extends JpaRepository<MessageVO, Integer>{
+
+}
