+++ src/main/java/com/foodietime/post/model/PostRepository.java
@@ -1,0 +1,10 @@
+package com.foodietime.post.model;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface PostRepository extends JpaRepository<PostVO, Integer>{
+
+}
